chapter_1_apache_beam/05_flex_template/requirements.txt
apache-beam
apache-beam[gcp]
google-cloud-storage
gcsfs
structlog